Movement and Library Setup
Added binary files for SDL2 and SDL2_image libraries.

Added base functionality for toad movement.

Added a SDL_Rect to the Toad struct for its texture's positioning.

Added width and height properties to the Screen struct.

Fixed a typo in the error message for a failed instance of loadTexture.

Restructured some code to allow for events to work properly in main.